[#2] Validate record field type

+++ src/test/resources/test_classes/missing_name/MissingNameSproc.java
@@ -9,7 +9,7 @@ import java.util.stream.Stream;
 public class MissingNameSproc {
 
     public static class GoodRecord {
-        public int age;
+        public long age;
     }
 
     @Context public GraphDatabaseService db;
